+++ src/docs/views/components/content/page-searchable-list/fragments/playground-searchable-list.vue
@@ -37,7 +37,7 @@ import { computed, ref } from "vue";
 import useTemplateGenerator from "@/docs/views/components/composables/use-template-generator/use-template-generator";
 
 // The current model value.
-const componentModel = ref(null);
+const componentModel = ref("");
 
 // Our base text slots, available for the user to update.
 const textSlots = ref({